Added index.css, projects.html, images folder. Updated Index.html
index.html:
Changed the nav bar to be pills instead of a drop down, added a div for the main body, made the footer have <p> to easily separate text, added a Last Updated page.
projects.html:
Created the projects.html file and copied the index.html code into it, then added a single card for the Egg Integrate Website, along with a button that adds more info for the project. Will add more cards in the future.
index.css:
I only added this so I could change the colors for the pills because dark blue on gray does not look well, and gray on gray is worse, so active links are now aqua and inactive links are now crimson.
images:
Added a png image of the Egg Integrate website.